requirements.txt
fastapi==0.112.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
sqladmin==0.18.0
sqlmodel==0.0.21
uvicorn[standard]==0.30.6
passlib==1.7.4
requests==2.32.3
pyjwt[crypto]==2.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==4.1.0
mysqlclient==2.2.4
Jinja2==3.1.2
gunicorn==23.0.0

#AÑADIDOS DE SANTIAGO PARA EL API DE REPORTES:
python-docx==1.1.2
docxtpl
plotly
pandas
kaleido==0.2.1
unidecode
openpyxl